requirements2.txt
azureml-sdk
azureml-widgets
azureml-defaults
matplotlib
seaborn
torch
torchvision
tensorboard
joblib
# git+git://github.com/TueLindhart/dtu_mlops_2021.git@azure#egg=src
# -e .